requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
python-docx>=1.1,<2.0
qrcode[pil]>=7.4,<8.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
python-multipart>=0.0.9,<1.0
